add cluter_id&location_id list option; list options goes to list_options.go
